pyproject.toml
[project]
name = "sqlquerypilot"
version = "0.1.0"
description = ""
authors = [
    {name = "chaminda satanarachchi",email = "chami.7896@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pyodbc (==5.1.0)",
    "tabulate (==0.9.0)",
    "openai (==1.56.1)",
    "langchain (==0.1.6)",
    "langchain-community (==0.0.20)",
    "langchain-core (==0.1.23)",
    "langchain-experimental (==0.0.49)",
    "langchain-openai (==0.0.5)",
    "pandas (==2.2.2)",
    "sqlalchemy (==2.0.30)",
    "pandas (==2.2.2)",
    "python-dotenv (==1.0.1)",
    "streamlit (>=1.42.2,<2.0.0)"
]

[tool.poetry]
packages = [{include = "sqlquerypilot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
